+++ comics/comics/menagea3.py
@@ -1,5 +1,7 @@
 # encoding: utf-8
 
+import datetime
+
 from comics.aggregator.crawler import CrawlerBase, CrawlerImage
 from comics.core.comic_data import ComicDataBase
 
@@ -18,10 +20,13 @@ class Crawler(CrawlerBase):
     time_zone = 'US/Eastern'
 
     def crawl(self, pub_date):
+        # Release to the feed is one day delayed, so we try to get yesterday's
+        # comic instead.
+        pub_date -= datetime.timedelta(days=1)
+
         feed = self.parse_feed('http://www.ma3comic.com/comic.rss')
         for entry in feed.for_date(pub_date):
             title = entry.title.replace('Menage a 3 - ', '')
-            page_url = entry.link.replace(' ', '+')
-            page = self.parse_page(page_url)
+            page = self.parse_page(entry.link)
             url = page.src('#cc img')
             return CrawlerImage(url, title)